[study] 수동 로또 - review 내용반영
