+++ src/main/java/service/LottoService.java
@@ -19,42 +19,21 @@ public class LottoService {
         int lottoCount = money / LottoPrice.PRICE.getPrice();
         InputView.printBuyCount(lottoCount);
 
-        lottos = new Lottos(makeLottoList(lottoCount));
+        lottos = new Lottos(lottoCount);
         ResultView.printLottoList(lottos.getLottos());
 
     }
 
     public LottoService(int money, int manualLottoCount) {
-        List<Lotto> newLottos = new ArrayList<>();
-
         ResultView.printManualLottoInput();
-
         int autoLottoCount = money / LottoPrice.PRICE.getPrice() - manualLottoCount;
-
-        IntStream.range(0, manualLottoCount).forEach(it -> {
-            String stringLotto = InputView.scanManualLotto();
-            newLottos.add(new Lotto(stringLotto));
-        });
-
-        newLottos.addAll(makeLottoList(autoLottoCount));
-
+        this.lottos = new Lottos(manualLottoCount, autoLottoCount);
         ResultView.printBuyManualAndAutoCount(manualLottoCount, autoLottoCount);
-        ResultView.printLottoList(newLottos);
-
-        this.lottos = new Lottos(newLottos);
+        ResultView.printLottoList(lottos.getLottos());
     }
 
     public void setFirstPlaceLotto(String input, String bonusLottoNumber) {
         firstPlaceLotto = new FirstPlaceLotto(input, bonusLottoNumber);
-    }
-
-    List<Lotto> makeLottoList(int lottoCount) {
-        List<Lotto> lottoList = new ArrayList<Lotto>();
-
-        IntStream.range(0, lottoCount)
-                .forEach(it -> lottoList.add(it, new Lotto()));
-
-        return lottoList;
     }
 
     public void printLottoResult() {
