+++ src/main/java/view/ResultView.java
@@ -33,14 +33,10 @@ public class ResultView {
         System.out.println("당첨 통계");
         System.out.println("---------");
         System.out.println("3개 일치 (5000원)- " + map.get(LottoRankResult.FIFTH) + "개");
-        System.out
-                .println("4개 일치 (50000원)- " + map.get(LottoRankResult.FOURTH) + "개");
-        System.out
-                .println("5개 일치 (1500000원)- " + map.get(LottoRankResult.THIRD) + "개");
-        System.out.println("5개 일치, 보너스 볼 일치(30000000원)- "
-                + map.get(LottoRankResult.SECOND) + "개");
-        System.out.println(
-                "6개 일치 (2000000000원)- " + map.get(LottoRankResult.FIRST) + "개");
+        System.out.println("4개 일치 (50000원)- " + map.get(LottoRankResult.FOURTH) + "개");
+        System.out.println("5개 일치 (1500000원)- " + map.get(LottoRankResult.THIRD) + "개");
+        System.out.println("5개 일치, 보너스 볼 일치(30000000원)- " + map.get(LottoRankResult.SECOND) + "개");
+        System.out.println("6개 일치 (2000000000원)- " + map.get(LottoRankResult.FIRST) + "개");
         System.out.println("총 수익률은 " + LottoProfit.getProfitRate(lottos, firstPlaceLotto) + "입니다.");
     }
